move runtime templates into standalone directory
