+++ README.md
@@ -47,9 +47,9 @@ generated by the compiler. While it is highly recommended, Typescript is NOT req
 houdini is available on npm:
 
 ```sh
-yarn add -D houdini houdini-preprocess graphql
+yarn add -D houdini houdini-preprocess
 # or
-npm install --save-dev houdini houdini-preprocess graphql
+npm install --save-dev houdini houdini-preprocess
 ```
 
 ## 🔧&nbsp;&nbsp;Configuring Your Application
